moved get items to calculator impl
